Solucion de errores y cambios.

+++ src/main/java/com/IngSoft/Evaluacion2/Costos/Costos.java
@@ -1,5 +1,5 @@
 package com.IngSoft.Evaluacion2.Costos;
 
 public interface Costos {
-    Integer getPrecio();
+    Long getPrecio();
 }
